requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-telegram-bot
docker
asyncio-mqtt
twisted
pyftpdlib
jinja2
plotly
python-multipart
httpx
scapy
aiohttp